refactor serialization tests a bit.
