+++ Tests/DataTests/AttributesTests.swift
@@ -19,18 +19,48 @@ class AttributesTests: XCTestCase {
         "testFalseBoolean": false,
         "nestedObject": ["anArray": [1, 2, 3, 4]]
     ]
+    
+    func verifyDecodedAttributes(attributes: Attributes) {
+        XCTAssertEqual(attributes.rawValue["testInt"] as! Int, 42)
+        XCTAssertEqual(attributes.rawValue["anArray"] as! [Int], [1, 2, 3, 4])
+        XCTAssertEqual(attributes.rawValue["testTrueBoolean"] as! Bool, true)
+        XCTAssertEqual(attributes.rawValue["testString"] as! String, "donut")
+        XCTAssertEqual(attributes.rawValue["testFalseBoolean"] as! Bool, false)
+        XCTAssertEqual((attributes.rawValue["nestedObject"] as! [String: Any])["anArray"] as! [Int], [1, 2, 3, 4])
+    }
 
     func testAttributesCoercion() {
         let dictionary = ["thing": 42]
         let attributes = dictionary.attributes
     }
 
-    func testCodableRoundtrip() {
+    func testCodableRoundtrip() throws {
+        // use JSONEncoder to test that Codable was synthesized properly.
+        let json = try JSONEncoder.default.encode(Attributes.init(exampleAttributes))
         
+        do {
+            let attributes = try JSONDecoder.default.decode(Attributes.self, from: json)
+            
+            verifyDecodedAttributes(attributes: attributes)
+        } catch {
+            // Print the error so the all-important UserInfo is captured:
+            print("Error decoding device: \(error)")
+            throw error
+        }
     }
     
-    func testNsCodingRoundtrip() {
-        
+    func testNSCodingRoundtrip() throws {
+        let archiver = NSKeyedArchiver.init(requiringSecureCoding: false)
+        let attributes = Attributes.init(exampleAttributes)
+        archiver.encodeRootObject(attributes)
+        //archiver.outputFormat = .xml
+        archiver.finishEncoding()
+        let archivedData = archiver.encodedData
+        let dearchiver = try NSKeyedUnarchiver.init(forReadingFrom: archivedData)
+        dearchiver.requiresSecureCoding = false
+        dearchiver.decodingFailurePolicy = .raiseException
+        let decodedAttributes = dearchiver.decodeObject() as! Attributes
+        verifyDecodedAttributes(attributes: decodedAttributes)
     }
     
 }
